# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/{userId}": {
      "get": {
        "operationId": "UsersController_getUserById",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user by ID",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/user-buckets": {
      "post": {
        "operationId": "UserBucketsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserBucketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User bucket created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBucket"
                }
              }
            }
          },
          "409": {
            "description": "User bucket relationship already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new user bucket relationship",
        "tags": [
          "User Buckets"
        ]
      },
      "get": {
        "operationId": "UserBucketsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of user buckets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserBucket"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all user buckets for the current user",
        "tags": [
          "User Buckets"
        ]
      }
    },
    "/api/v1/user-buckets/{id}": {
      "get": {
        "operationId": "UserBucketsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User bucket found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBucket"
                }
              }
            }
          },
          "404": {
            "description": "User bucket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific user bucket by ID",
        "tags": [
          "User Buckets"
        ]
      },
      "patch": {
        "operationId": "UserBucketsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserBucketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User bucket updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBucket"
                }
              }
            }
          },
          "404": {
            "description": "User bucket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a user bucket",
        "tags": [
          "User Buckets"
        ]
      },
      "delete": {
        "operationId": "UserBucketsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User bucket deleted successfully"
          },
          "404": {
            "description": "User bucket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a user bucket",
        "tags": [
          "User Buckets"
        ]
      }
    },
    "/api/v1/user-buckets/bucket/{bucketId}/snooze-status": {
      "get": {
        "operationId": "UserBucketsController_getSnoozeStatus",
        "parameters": [
          {
            "name": "bucketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snooze status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Check if a bucket is snoozed for the current user",
        "tags": [
          "User Buckets"
        ]
      }
    },
    "/api/v1/user-buckets/bucket/{bucketId}/snooze": {
      "post": {
        "operationId": "UserBucketsController_setBucketSnooze",
        "parameters": [
          {
            "name": "bucketId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bucket snooze set successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBucket"
                }
              }
            }
          },
          "201": {
            "description": "Bucket snooze created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBucket"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Set bucket snooze for the current user",
        "tags": [
          "User Buckets"
        ]
      }
    },
    "/api/v1/user-buckets/snoozed-bucket-ids": {
      "get": {
        "operationId": "UserBucketsController_getSnoozedBucketIds",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of snoozed bucket IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all snoozed bucket IDs for the current user",
        "tags": [
          "User Buckets"
        ]
      }
    },
...
    "/api/v1/buckets/{id}": {
      "get": {
        "operationId": "BucketsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bucket details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bucket"
                }
              }
            }
          },
          "404": {
            "description": "Bucket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a specific bucket by ID",
        "tags": [
          "Buckets"
        ]
      },
      "patch": {
        "operationId": "BucketsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBucketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bucket updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Bucket"
                }
              }
            }
          },
          "404": {
            "description": "Bucket not found or one or more devices not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a bucket and its device associations",
        "tags": [
          "Buckets"
        ]
      },
      "delete": {
        "operationId": "BucketsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bucket deleted successfully"
          },
          "404": {
            "description": "Bucket not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a bucket",
        "tags": [
          "Buckets"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zentik API",
    "description": "The Zentik notification API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-string"
          },
          "email": {
            "type": "string",
            "example": "john@example.com"
          },
          "username": {
            "type": "string",
            "example": "john_doe"
          },
          "hasPassword": {
            "type": "boolean",
            "example": true,
            "description": "Whether the user has a password set"
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "avatar": {
            "type": "string",
            "example": "https://example.com/avatar.jpg"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "moderator",
              "admin"
            ],
            "default": "user"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "resetToken": {
            "type": "object"
          },
          "resetTokenRequestedAt": {
            "type": "object"
          },
          "emailConfirmationToken": {
            "type": "object"
          },
          "emailConfirmationTokenRequestedAt": {
            "type": "object"
          },
          "emailConfirmed": {
            "type": "boolean",
            "example": false,
            "description": "Whether the user email is confirmed"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "hasPassword",
          "role",
          "createdAt",
          "updatedAt",
          "emailConfirmed"
        ]
      },
...
      "SnoozeSchedule": {
        "type": "object",
        "properties": {
          "days": {
            "description": "Days of the week when snooze is active",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeFrom": {
            "type": "string",
            "description": "Start time for snooze period"
          },
          "timeTill": {
            "type": "string",
            "description": "End time for snooze period"
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether this snooze schedule is enabled"
          }
        },
        "required": [
          "days",
          "timeFrom",
          "timeTill",
          "isEnabled"
        ]
      },
      "UserBucket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "bucket": {
            "$ref": "#/components/schemas/Bucket"
          },
          "bucketId": {
            "type": "string"
          },
          "snoozeUntil": {
            "type": "object",
            "description": "Date until which notifications from this bucket are snoozed"
          },
          "snoozes": {
            "description": "Array of snooze schedules for recurring snooze periods",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnoozeSchedule"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "userId",
          "bucket",
          "bucketId",
          "createdAt",
          "updatedAt"
        ]
      },
      "EntityPermission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "resourceType": {
            "type": "string",
            "description": "Type of the resource (e.g., bucket, notification, user_webhook)"
          },
          "resourceId": {
            "type": "string",
            "description": "ID of the resource"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "grantedBy": {
            "description": "User who granted these permissions",
            "allOf": [
              {
                "$ref": "#/components/schemas/User"
              }
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "read",
                "write",
                "delete",
                "admin"
              ]
            }
          },
          "expiresAt": {
            "format": "date-time",
            "type": "string",
            "description": "Optional expiration date for the permissions"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "resourceType",
          "resourceId",
          "user",
          "permissions",
          "createdAt",
          "updatedAt"
        ]
      },
      "Bucket": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for the bucket (e.g., #FF5733)"
          },
          "isProtected": {
            "type": "boolean",
            "description": "Whether the bucket is protected from deletion",
            "default": false
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether the bucket is publicly accessible",
            "default": false
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "userBuckets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserBucket"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EntityPermission"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "user",
          "messages",
          "createdAt",
          "updatedAt"
        ]
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "subtitle": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "attachments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "fileAttachments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "attachmentUuids": {
            "description": "List of attachment UUIDs for this message",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "tapAction": {
            "type": "object"
          },
          "sound": {
            "type": "string"
          },
          "deliveryType": {
            "type": "string",
            "enum": [
              "SILENT",
              "NORMAL",
              "CRITICAL"
            ]
          },
          "addMarkAsReadAction": {
            "type": "boolean"
          },
          "addOpenNotificationAction": {
            "type": "boolean"
          },
          "addDeleteAction": {
            "type": "boolean"
          },
          "snoozes": {
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "locale": {
            "type": "string"
          },
          "bucket": {
            "$ref": "#/components/schemas/Bucket"
          },
          "bucketId": {
            "type": "string"
          },
          "groupId": {
            "type": "string",
            "description": "Optional group ID for notification grouping, falls back to bucketId if not provided"
          },
          "collapseId": {
            "type": "string",
            "description": "Optional collapse ID for APNS collapse-id, used to replace notifications with the same collapse ID"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "deliveryType",
          "bucket",
          "bucketId",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "SnoozeScheduleInput": {
        "type": "object",
        "properties": {
          "days": {
            "description": "Days of the week when snooze is active",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "timeFrom": {
            "type": "string",
            "description": "Start time for snooze period"
          },
          "timeTill": {
            "type": "string",
            "description": "End time for snooze period"
          },
          "isEnabled": {
            "type": "boolean",
            "description": "Whether this snooze schedule is enabled"
          }
        },
        "required": [
          "days",
          "timeFrom",
          "timeTill",
          "isEnabled"
        ]
      },
      "CreateUserBucketDto": {
        "type": "object",
        "properties": {
          "bucketId": {
            "type": "string",
            "description": "ID of the bucket"
          },
          "snoozeUntil": {
            "type": "string",
            "description": "Date until which notifications from this bucket are snoozed"
          },
          "snoozes": {
            "description": "Array of snooze schedules for recurring snooze periods",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SnoozeScheduleInput"
            }
          }
        },
        "required": [
          "bucketId"
        ]
      },
      "UpdateUserBucketDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateBucketDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the bucket"
          },
          "icon": {
            "type": "string",
            "description": "Icon for the bucket"
          },
          "description": {
            "type": "string",
            "description": "Description of the bucket"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for the bucket (e.g., #FF5733)",
            "pattern": "^#[0-9A-Fa-f]{6}$",
            "example": "#0a7ea4"
          },
          "isProtected": {
            "type": "boolean",
            "description": "Whether the bucket is protected from deletion",
            "default": false
          },
          "isPublic": {
            "type": "boolean",
            "description": "Whether the bucket is publicly accessible",
            "default": false
          }
        }
      },
...
    }
  }
}